basic refactoring to make use of Rooms more broadly and put the model definition into its own file; display the pubsno, if it exists; show the con name on the printed pages
