+++ webpages/schedule_table_renderer.php
@@ -1,7 +1,6 @@
 <?php
 
-define('__ROOT__', dirname(__FILE__)); 
-require_once(__ROOT__.'/con_data.php');
+require_once(__DIR__ . '/con_data.php');
 
 function determine_con_start_date() {
     $conData = ConData::fromEnvironmentDefinition();
@@ -44,21 +43,21 @@ class ScheduleTableRenderer {
             $height = count($headerRows) - $rowNumber - $value->getRowHeight() + 1;
             $rowHeight = $height == 1 ? "" : "rowspan=\"$height\"";
             if ($this->showRoomArea) {
-                $header .= "<th $rowHeight $width>" . $value->roomName 
-                    . ($value->isOnline ? " <span class=\"small\"><br />(Online)</span>" : "") 
-                    . ($value->area ? ("<span class=\"small\"><br />" . number_format($value->area) . " sq ft</span>") : "") 
+                $header .= "<th $rowHeight $width>" . $value->roomName
+                    . ($value->isOnline ? " <span class=\"small\"><br />(Online)</span>" : "")
+                    . ($value->area ? ("<span class=\"small\"><br />" . number_format($value->area) . " sq ft</span>") : "")
                     . "</th>";
             } else {
                 $header .= "<th $rowHeight $width>" . $value->roomName . "</th>";
             }
-    
+
             if ($value->children && count($value->children) > 0) {
                 $this->buildTableHeaderRows($headerRows, $value->children, $rowNumber + 1);
             }
         }
         $headerRows[$rowNumber] = $header;
     }
-    
+
     function renderTableHeader() {
         $maxRows = 1;
         foreach ($this->rooms as $r) {
@@ -101,10 +100,10 @@ EOD;
         </ul>
         <div class="tab-content">
 EOD;
-    
+
         $lastRoom = $this->rooms[count($this->rooms)-1];
         $maxColumns = $lastRoom->columnNumber + $lastRoom->getColumnWidth() - 1;
-    
+
         $startDate = determine_con_start_date();
         for ($day = 0; $day < CON_NUM_DAYS; $day++) {
             $classes = ($day === 0) ? " show active" : "";
@@ -119,11 +118,11 @@ EOD;
 
             echo "<tbody>";
             echo "<tr><th colspan=\"" . ($maxColumns + 2) . "\">" . $startDate->format('D, d M') . "</th></tr>";
-            
+
             if ($this->dataProvider->isCellDataAvailableForDay($day)) {
                 $fromIndex = $this->dataProvider->findFirstStartIndexForDay($day);
                 $toIndex = $this->dataProvider->findLastEndIndexForDay($day);
-    
+
                 for ($row = $fromIndex; $row <= $toIndex; $row++) {
                     echo "<tr>";
                     echo "<th class=\"bg-light small\">";
@@ -133,13 +132,13 @@ EOD;
                             $hours -= 24;
                         }
                         $minutes = str_pad(($row % 4) * 15, 2, "0", STR_PAD_LEFT);
-    
+
                         echo "$hours:$minutes";
                     } else {
                         echo "&nbsp;";
                     }
                     echo "</th>";
-    
+
                     for ($column = 0; $column <= $maxColumns; $column++) {
                         $slot = $this->dataProvider->getCellDataFor($row, $column, $day);
                         if ($slot) {
